python/fastapi-sse-ex/pyproject.toml
[tool.poetry]
name = "fastapi-sse-ex"
version = "0.1.0"
description = ""
authors = ["daengdaengLee <gunho1020@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
openai = "^1.9.0"
fastapi = {extras = ["all"], version = "^0.109.0"}


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
